fixed another baum-welch
Fixed an error with Model.log_probability, where the wrong indices where being used. Int j was supposed to iterate through the entire path, but would iterate through the path up until the length of the sequence instead. Also, pair_lse was the wrong approach, should be simple addition.

Made a change to Baum-Welch which will allow for summary statistics in the future-- in that instead of using a slice of the emission weights matrix for training, it will go through this matrix, determine which symbols are paired with which states, and append the symbol and weight to a growing list for that state. A modification to the tied state modification was made, such that instead of changing all the entries in the emission matrix, just gather the correct probability and use that, leaving the underlying matrix untouched.