minor change to maximum a posteriori
Maximum a Posteriori performs the calculation of the emission weights inside the function, instead of calling Model.forward_backward in order to save time from calculating the transition matrix, which is not used. However, in writing it, tied states were included. This will throw off the ML path, as any state in the set of tied states will be equally likely according to the algorithm. Tied states are only useful in training, and so tied states are not considered in MAP, as it should be.