viterbi index reported changed
Previously, the viterbi path would return ( log_probability, path ), where path is made up of ( index, state_object ) tuples. The index used to just be the index of the path, aka a 7 state path would have indexes from 0 to 7. This is useless, as you can just enumerate the path. This has been changes so the index is the index of the state object in the self.states list, which is an actually useful number.