transition matrix, distributions, fb
Dense transition matrix return now by Model.dense_transition_matrix().
This may not always be a good idea, but it's worth having.

Distribution string format changed a bit to make it work. Mixture
distributions were actually broken when writing and then reading in
again due to an error in the log_probability method, and the __str__
method. These have been amended.

In the forward_backward algorithm, tie=False has been toggled. This is a
toggle which allows you to return emission probabilities in a tied
manner, which does not cause the rows to sum to 1. As this feature is
not the normal one, I made it default to false.