added nanopore data example
ProfileHMM.ipynb is an IPython notebook which shows how to make HMM
modules for Profile HMMs using PyPore with an underlying YAHMM model.
HMM modules allow for forked profiles to be made, in addition to linear
HMMs, which can allow for classification of a sequence to be performed
in addition to just alignment to a reference.