expanding unit testing for distributions
Added a unit test for each of the distributions. Some are more in depth
than others, but at least one test is done per distribution, to ensure
that modifications to the code do not effect the log probabilities
generated.

Also, Uniform Kernel Density was calculating incorrect scores, and so
has been updated to be correct. All Kernel Densities wouldn't update
their parameters on training--this has been fixed.