adding batch summarizing for training
When performing Baum-Welch training, instead of passing all observations from all sequences in at a time, instead there is the option to send one sequence at a time and save a summary statistic, and compile the summary statistics into an update at the end. Currently in progress, as Uniform, Normal, LogNormal, Exponential, Discrete distributions have been updated but not the others.