Model.add_transitions added
Model.add_transitions allows you to add multiple transitions at the same
time in many ways. From the docstring:

(1) If both a and b are lists, then create transitions from the i-th
element of a to the i-th element of b with a probability equal to the
i-th element of probabilities.

Example:
model.add_transitions([model.start, s1], [s1, model.end], [1., 1.])

(2) If either a or b are a state, and the other is a list, create a
transition from all states in the list to the single state object with
probabilities and pseudocounts specified appropriately.

Example:
model.add_transitions([model.start, s1, s2, s3], s4, [0.2, 0.4, 0.3,
0.9])
model.add_transitions(model.start, [s1, s2, s3], [0.6, 0.2, 0.05])