v1.1.1
(1) log_probability( model, samples ) function added, which calculates
the log probability of a group of samples given a model. This can be
either a distribution or a HMM.

(2) MixtureDistribution bug fixed in which it would only look at the
first two distributions when calculating log probability

(3) EM algorithm added to train MixtureDistributions. Both emission and
weight updates added.

(4) Error in NormalDistribution.from_summaries in which it did not
respect min_std. This worked for from_sample but not from_summaries.